Fixed data_logger comparison None vs dict
- if a const_param was None, but some experiments had that key as a
dictionary, we still want to get all of the experiments with that key
set to None. With previous implementation would throw error if any value
loaded with that key was a dict. Instead, a check is done if the
const_param is a dict to throw the error. If the key passes that test
and the loaded key value is a dict, it is omitted since it's different